Improve desk hiding functionality

+++ app/src/main/java/com/meteoalgerie/autoscan/common/util/ViewUtil.kt
@@ -7,13 +7,9 @@ import androidx.annotation.AttrRes
 import androidx.annotation.ColorInt
 import androidx.annotation.Dimension
 
-fun dpToPx(context: Context, @Dimension(unit = Dimension.DP) dp: Int): Int {
+fun dpToPx(context: Context, @Dimension(unit = Dimension.DP) dp: Float): Int {
     val r = context.resources
-    return TypedValue.applyDimension(
-        TypedValue.COMPLEX_UNIT_DIP,
-        dp.toFloat(),
-        r.displayMetrics
-    ).toInt()
+    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, r.displayMetrics).toInt()
 }
 
 @ColorInt
